+++ lib/ethers/types.ex
@@ -13,6 +13,22 @@ defmodule Ethers.Types do
   @type t_address :: <<_::336>> | <<_::160>>
 
   @typedoc """
+  Public key either in its uncompressed format (64 bytes MAY BE prefixed with 0x04) or its
+  compressed format (32 bytes MUST BE prefixed with 0x02 or 0x03)
+
+  It can be hex encoded but only with 0x prefix.
+
+  Compressed public key MUST have a prefix.
+  """
+  @type t_pub_key ::
+          <<_::520>>
+          | <<_::512>>
+          | <<_::264>>
+          | <<_::1042>>
+          | <<_::1026>>
+          | <<_::530>>
+
+  @typedoc """
   keccak hash in its hex format with 0x
 
   ## Examples